pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=60",
  "setuptools_scm[toml]>=8.0"
]

[project]
authors = [
  {name="Ryan Roussel"},
  {name="Juan Pablo Gonzalez Aguilera"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "lightning",
  "numpy",
  "matplotlib",
  "pytorch",
  "pytorch-cuda",
  "cheetah-accelerator"
]
description = "Phase space reconstruction using differentiable beam dynamics and generative beam representations"
dynamic = [ "version" ]
keywords = []
name = "gpsr"
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
  "pytest",
  "ffmpeg",
  "pytest",
  "pytest-cov",
  "jupyterlab>=3",
  "jupyterlab-lsp",
  "python-lsp-server",
  "pygments",
]
doc = [
  "mkdocs",
  "mkdocs-jupyter",
  "mkdocs-macros-plugin",
  "mkdocs-material",
  "mkdocstrings",
  "mkdocstrings-python",
]

[project.urls]
Homepage = "https://github.com/roussel-ryan/gpsr"

[options]
zip_safe = false
include_package_data = true

[tool.setuptools.packages.find]
where = ["."]
include = [ "gpsr", ]
namespaces = false

[project.license]
file = "LICENSE"

[tool.ruff]
# select = []
# ignore = []

[tool.pytest.ini_options]
addopts = "--cov=gpsr/"
log_cli_level = "info"
log_level = "debug"
testpaths = ["gpsr/tests"]

[tool.setuptools_scm]
version_file = "gpsr/_version.py"
